Put format utils to CoreRequest
